+++ src/fpnew_cast_multi.sv
@@ -199,11 +199,11 @@ module fpnew_cast_multi #(
       // Compensation for the difference in mantissa widths used for leading-zero count
       assign fmt_shift_compensation[fmt] = signed'(INT_MAN_WIDTH - 1 - MAN_BITS);
     end else begin : inactive_format
-      assign info_q[fmt]                 = 'X; // propagate don't care (format disabled)
-      assign fmt_sign[fmt]               = 'x; // propagate don't care (format disabled)
-      assign fmt_exponent[fmt]           = 'X; // propagate don't care (format disabled)
-      assign fmt_mantissa[fmt]           = 'X; // propagate don't care (format disabled)
-      assign fmt_shift_compensation[fmt] = 'X; // propagate don't care (format disabled)
+      assign info_q[fmt]                 = '{default: fpnew_pkg::DONT_CARE}; // format disabled
+      assign fmt_sign[fmt]               = fpnew_pkg::DONT_CARE;             // format disabled
+      assign fmt_exponent[fmt]           = '{default: fpnew_pkg::DONT_CARE}; // format disabled
+      assign fmt_mantissa[fmt]           = '{default: fpnew_pkg::DONT_CARE}; // format disabled
+      assign fmt_shift_compensation[fmt] = '{default: fpnew_pkg::DONT_CARE}; // format disabled
     end
   end
 
@@ -219,7 +219,7 @@ module fpnew_cast_multi #(
         ifmt_input_val[ifmt][INT_WIDTH-1:0] = operands_q[INT_WIDTH-1:0];
       end
     end else begin : inactive_format
-      assign ifmt_input_val[ifmt] = 'X; // don't care about disabled formats
+      assign ifmt_input_val[ifmt] = '{default: fpnew_pkg::DONT_CARE}; // format disabled
     end
   end
 
@@ -492,7 +492,7 @@ module fpnew_cast_multi #(
         fmt_pre_round_abs[fmt] = {final_exp[EXP_BITS-1:0], final_mant[MAN_BITS-1:0]}; // 0-extend
       end
     end else begin : inactive_format
-      assign fmt_pre_round_abs[fmt] = 'X;
+      assign fmt_pre_round_abs[fmt] = '{default: fpnew_pkg::DONT_CARE};
     end
   end
 
@@ -508,7 +508,7 @@ module fpnew_cast_multi #(
         ifmt_pre_round_abs[ifmt][INT_WIDTH-1:0] = final_int[INT_WIDTH-1:0];
       end
     end else begin : inactive_format
-      assign ifmt_pre_round_abs[ifmt] = 'X;
+      assign ifmt_pre_round_abs[ifmt] = '{default: fpnew_pkg::DONT_CARE};
     end
   end
 
@@ -550,9 +550,9 @@ module fpnew_cast_multi #(
                                         : {rounded_sign, rounded_abs[EXP_BITS+MAN_BITS-1:0]};
       end
     end else begin : inactive_format
-      assign fmt_uf_after_round[fmt] = 'X;
-      assign fmt_of_after_round[fmt] = 'X;
-      assign fmt_result[fmt]         = 'X;
+      assign fmt_uf_after_round[fmt] = fpnew_pkg::DONT_CARE;
+      assign fmt_of_after_round[fmt] = fpnew_pkg::DONT_CARE;
+      assign fmt_result[fmt]         = '{default: fpnew_pkg::DONT_CARE};
     end
   end
 
@@ -596,7 +596,7 @@ module fpnew_cast_multi #(
         fmt_special_result[fmt][FP_WIDTH-1:0] = special_res;
       end
     end else begin : inactive_format
-      assign fmt_special_result[fmt] = 'X;
+      assign fmt_special_result[fmt] = '{default: fpnew_pkg::DONT_CARE};
     end
   end
 
@@ -642,7 +642,7 @@ module fpnew_cast_multi #(
         ifmt_special_result[ifmt][INT_WIDTH-1:0] = special_res;
       end
     end else begin : inactive_format
-      assign ifmt_special_result[ifmt] = 'X;
+      assign ifmt_special_result[ifmt] = '{default: fpnew_pkg::DONT_CARE};
     end
   end