+++ src/fpnew_noncomp.sv
@@ -187,7 +187,7 @@ module fpnew_noncomp #(
       fpnew_pkg::RTZ: sgnj_result.sign = ~sign_b;         // SGNJN
       fpnew_pkg::RDN: sgnj_result.sign = sign_a ^ sign_b; // SGNJX
       fpnew_pkg::RUP: sgnj_result      = operand_a;       // passthrough
-      default: sgnj_result = 'X; // propagate X
+      default: sgnj_result = '{default: fpnew_pkg::DONT_CARE}; // don't care
     endcase
   end
 
@@ -223,7 +223,7 @@ module fpnew_noncomp #(
       unique case (rnd_mode_q)
         fpnew_pkg::RNE: minmax_result = operand_a_smaller ? operand_a : operand_b; // MIN
         fpnew_pkg::RTZ: minmax_result = operand_a_smaller ? operand_b : operand_a; // MAX
-        default: minmax_result = 'X; // propagate X
+        default: minmax_result = '{default: fpnew_pkg::DONT_CARE}; // don't care
       endcase
     end
   end
@@ -262,7 +262,7 @@ module fpnew_noncomp #(
           if (any_operand_nan) cmp_result = op_mod_q; // NaNs are valid, always campare as not equal
           else cmp_result = operands_equal ^ op_mod_q;
         end
-        default: cmp_result = 'X; // propagate X
+        default: cmp_result = '{default: fpnew_pkg::DONT_CARE}; // don't care
       endcase
     end
   end
@@ -323,14 +323,14 @@ module fpnew_noncomp #(
         extension_bit_d = cmp_extension_bit;
       end
       fpnew_pkg::CLASSIFY: begin
-        result_d        = 'X; // unused
+        result_d        = '{default: fpnew_pkg::DONT_CARE}; // unused
         status_d        = class_status;
         extension_bit_d = class_extension_bit;
       end
       default: begin
-        result_d        = 'X; // propaagate X
-        status_d        = 'X; // propaagate X
-        extension_bit_d = 'X; // propaagate X
+        result_d        = '{default: fpnew_pkg::DONT_CARE}; // dont care
+        status_d        = '{default: fpnew_pkg::DONT_CARE}; // dont care
+        extension_bit_d = fpnew_pkg::DONT_CARE;             // dont care
       end
     endcase
   end